backend/requirements.txt
setuptools>=70,<81
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
crewai==0.80.0
# crewai-tools removido — não usado e incompatível
anthropic==0.49.0
openai==1.58.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pydantic==2.9.0
litellm==1.55.0
httpx==0.27.0
sqlalchemy==2.0.36
cryptography>=43.0.3,<44.0.0
apscheduler==3.10.4
